test(SDK-5354) - Makes classes smaller
