+++ clevertap-core/src/main/java/com/clevertap/android/sdk/events/EventQueueManager.java
@@ -25,7 +25,7 @@ import com.clevertap.android.sdk.login.IdentityRepo;
 import com.clevertap.android.sdk.login.IdentityRepoFactory;
 import com.clevertap.android.sdk.login.LoginInfoProvider;
 import com.clevertap.android.sdk.network.NetworkManager;
-import com.clevertap.android.sdk.profile.ProfileStateMerger;
+import com.clevertap.android.sdk.profile.merge.ProfileChange;
 import com.clevertap.android.sdk.task.CTExecutorFactory;
 import com.clevertap.android.sdk.task.MainLooperHandler;
 import com.clevertap.android.sdk.task.Task;
@@ -329,7 +329,7 @@ public class EventQueueManager extends BaseEventQueueManager implements FailureF
             Map<String, Object> flattenedEventProps = ((FlattenedEventData.EventProperties) flattenedEventData).getProperties();
             controllerManager.getInAppController().onQueueEvent(eventName, flattenedEventProps, userLocation);
         } else if (flattenedEventData instanceof FlattenedEventData.ProfileChanges) {
-            Map<String, ProfileStateMerger.ProfileChange> flattenedProfileChanges = ((FlattenedEventData.ProfileChanges) flattenedEventData).getChanges();
+            Map<String, ProfileChange> flattenedProfileChanges = ((FlattenedEventData.ProfileChanges) flattenedEventData).getChanges();
             controllerManager.getInAppController().onQueueProfileEvent(flattenedProfileChanges, userLocation);
         } else if (!eventMediator.isAppLaunchedEvent(event) && flattenedEventData instanceof FlattenedEventData.EventProperties) {
             // in case device is online only evaluate non-appLaunched events
