+++ clevertap-core/src/main/java/com/clevertap/android/sdk/events/FlattenedEventData.kt
@@ -1,9 +1,9 @@
 package com.clevertap.android.sdk.events
 
-import com.clevertap.android.sdk.profile.ProfileStateMerger
+import com.clevertap.android.sdk.profile.merge.ProfileChange
 
 sealed interface FlattenedEventData {
-    data class ProfileChanges(val changes: Map<String, ProfileStateMerger.ProfileChange>) : FlattenedEventData
+    data class ProfileChanges(val changes: Map<String, ProfileChange>) : FlattenedEventData
     data class EventProperties(val properties: Map<String, Any>) : FlattenedEventData
     data object NoData : FlattenedEventData
 }
