+++ clevertap-core/src/main/java/com/clevertap/android/sdk/profile/ProfileCommand.kt
@@ -1,5 +1,6 @@
 package com.clevertap.android.sdk.profile
 import com.clevertap.android.sdk.Constants
+import com.clevertap.android.sdk.profile.merge.MergeOperation
 
 /**
  * Maps profile command strings to their corresponding merge operations.
@@ -7,30 +8,16 @@ import com.clevertap.android.sdk.Constants
  * and the internal ProfileStateMerger operations.
  *
  * @property commandString The command string used in the API (e.g., "$add", "$incr")
- * @property operation The corresponding ProfileStateMerger.MergeOperation
+ * @property operation The corresponding MergeOperation
  */
 internal enum class ProfileCommand(
     val commandString: String,
-    val operation: ProfileStateMerger.MergeOperation
+    val operation: MergeOperation
 ) {
-    SET(Constants.COMMAND_SET, ProfileStateMerger.MergeOperation.UPDATE),
-    ADD(Constants.COMMAND_ADD, ProfileStateMerger.MergeOperation.ARRAY_ADD),
-    REMOVE(Constants.COMMAND_REMOVE, ProfileStateMerger.MergeOperation.ARRAY_REMOVE),
-    DELETE(Constants.COMMAND_DELETE, ProfileStateMerger.MergeOperation.DELETE),
-    INCREMENT(Constants.COMMAND_INCREMENT, ProfileStateMerger.MergeOperation.INCREMENT),
-    DECREMENT(Constants.COMMAND_DECREMENT, ProfileStateMerger.MergeOperation.DECREMENT);
-
-    companion object {
-        /**
-         * Converts a command string to its corresponding MergeOperation.
-         *
-         * @param command The command string (e.g., "$add", "$incr")
-         * @return The corresponding MergeOperation, or null if not found
-         */
-        @JvmStatic
-        fun fromCommand(command: String?): ProfileStateMerger.MergeOperation? {
-            if (command == null) return null
-            return values().firstOrNull { it.commandString == command }?.operation
-        }
-    }
+    SET(Constants.COMMAND_SET, MergeOperation.UPDATE),
+    ADD(Constants.COMMAND_ADD, MergeOperation.ARRAY_ADD),
+    REMOVE(Constants.COMMAND_REMOVE, MergeOperation.ARRAY_REMOVE),
+    DELETE(Constants.COMMAND_DELETE, MergeOperation.DELETE),
+    INCREMENT(Constants.COMMAND_INCREMENT, MergeOperation.INCREMENT),
+    DECREMENT(Constants.COMMAND_DECREMENT, MergeOperation.DECREMENT);
 }
