+++ clevertap-core/src/main/java/com/clevertap/android/sdk/profile/ProfileStateMerger.kt
@@ -1,7 +1,12 @@
 package com.clevertap.android.sdk.profile
 
 import androidx.annotation.WorkerThread
-import org.json.JSONArray
+import com.clevertap.android.sdk.profile.merge.ArrayOperationHandler
+import com.clevertap.android.sdk.profile.merge.DeleteOperationHandler
+import com.clevertap.android.sdk.profile.merge.MergeOperation
+import com.clevertap.android.sdk.profile.merge.ProfileChange
+import com.clevertap.android.sdk.profile.merge.ProfileChangeTracker
+import com.clevertap.android.sdk.profile.merge.UpdateOperationHandler
 import org.json.JSONException
 import org.json.JSONObject
 
@@ -18,69 +23,20 @@ import org.json.JSONObject
  * @see MergeOperation for available operations
  * @see ProfileChange for tracking changes
  */
-// todo check for JSONObject.of
-// todo changes for $D_ should be reported as without $D
 class ProfileStateMerger {
 
-    /**
-     * Defines the type of merge operation to perform on profile data.
-     */
-    enum class MergeOperation {
-        /**
-         * Replace old values with new values. Creates keys if they don't exist.
-         * For nested objects, recursively merges properties.
-         */
-        UPDATE,
+    private val changeTracker = ProfileChangeTracker()
+    private val arrayHandler = ArrayOperationHandler()
+    private val updateHandler = UpdateOperationHandler(changeTracker, arrayHandler)
+    private val deleteHandler = DeleteOperationHandler(changeTracker)
 
-        /**
-         * Add new value to old value (numeric values only).
-         * Skips non-numeric values and keys that don't exist.
-         */
-        INCREMENT,
-
-        /**
-         * Subtract new value from old value (numeric values only).
-         * Skips non-numeric values and keys that don't exist.
-         */
-        DECREMENT,
-
-        /**
-         * Delete keys specified in source from target.
-         * Use DELETE_MARKER constant as value to mark fields for deletion.
-         */
-        DELETE,
-
-        /**
-         * Add string values to arrays.
-         * Creates array if it doesn't exist. Skips duplicates.
-         * Only works with string array elements.
-         */
-        ARRAY_ADD,
-
-        /**
-         * Remove string values from arrays.
-         * Only works with string array elements.
-         */
-        ARRAY_REMOVE
-    }
-
-    /**
-     * Represents a change in profile state, tracking old and new values.
-     *
-     * @property oldValue The previous value (null if newly added)
-     * @property newValue The new value (null if deleted)
-     */
-    data class ProfileChange(
-        val oldValue: Any?,
-        val newValue: Any?
-    )
     /**
      * Result of a merge operation containing all changes made.
      *
      * @property changes Map of dot-notation paths to ProfileChange objects
      */
     data class MergeResult(
-        val changes: Map<String, ProfileChange>,
+        val changes: Map<String, ProfileChange>
     )
 
     /**
@@ -130,497 +86,23 @@ class ProfileStateMerger {
             val currentPath = buildPath(path, key)
 
             when (operation) {
-                MergeOperation.DELETE -> handleDelete(target, key, newValue, currentPath, changes)
-                else -> handleMerge(target, key, newValue, currentPath, changes, operation)
-            }
-        }
-    }
-
-    /**
-     * Handles deletion operations.
-     */
-    private fun handleDelete(
-        target: JSONObject,
-        key: String,
-        newValue: Any,
-        currentPath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        if (!target.has(key)) return
-
-        val oldValue = target.get(key)
-
-        when {
-            isDeleteMarker(newValue) -> {
-                // Delete this key entirely
-                deleteValue(target, key, oldValue, currentPath, changes)
-            }
-            oldValue is JSONObject && newValue is JSONObject -> {
-                // Recurse into nested objects
-                mergeRecursive(oldValue, newValue, currentPath, changes, MergeOperation.DELETE)
-                // Remove the object if it's now empty
-                if (oldValue.length() == 0) {
-                    target.remove(key)
-                }
-            }
-            oldValue is JSONArray && newValue is JSONArray -> {
-                // Handle array element deletions
-                handleArrayDeletion(target, key, oldValue, newValue, currentPath, changes)
-            }
-        }
-    }
-
-    /**
-     * Handles array deletions (element removal or field deletion from array elements).
-     */
-    private fun handleArrayDeletion(
-        parentJson: JSONObject,
-        key: String,
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        currentPath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        if (newArray.length() == 0) return
-
-        val hasDeleteMarkers = hasDeleteMarkerElements(newArray)
-        val hasObjectsToDelete = hasJsonObjectElements(newArray)
-
-        when {
-            hasDeleteMarkers -> deleteArrayElements(oldArray, newArray, currentPath, changes)
-            hasObjectsToDelete -> deleteFromArrayElements(oldArray, newArray, currentPath, changes)
-            else -> deleteValue(parentJson, key, oldArray, currentPath, changes)
-        }
-    }
-
-    /**
-     * Deletes specific fields from array elements.
-     */
-    private fun deleteFromArrayElements(
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        basePath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val oldArrayCopy = copyArray(oldArray)
-        var arrayModified = false
-
-        for (i in 0 until minOf(newArray.length(), oldArray.length())) {
-            val oldElement = oldArray.get(i)
-            val newElement = newArray.get(i)
-
-            if (oldElement is JSONObject && newElement is JSONObject) {
-                mergeRecursive(oldElement, newElement, "", mutableMapOf(), MergeOperation.DELETE)
-                arrayModified = true
-
-                // Remove empty objects
-                if (oldElement.length() == 0) {
-                    oldArray.remove(i)
-                }
-            }
-        }
-
-        if (arrayModified) {
-            changes[basePath] = ProfileChange(oldArrayCopy, oldArray)
-        }
-    }
-
-    /**
-     * Deletes entire array elements at specific indices marked with DELETE_MARKER.
-     */
-    private fun deleteArrayElements(
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        basePath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val oldArrayCopy = copyArray(oldArray)
-        val indicesToDelete = mutableListOf<Int>()
-
-        // Collect indices to delete
-        for (i in 0 until newArray.length()) {
-            if (isDeleteMarker(newArray.opt(i)) && i < oldArray.length()) {
-                indicesToDelete.add(i)
-            }
-        }
-
-        if (indicesToDelete.isEmpty()) return
-
-        // Delete in reverse order to maintain correct indices
-        indicesToDelete.sortedDescending().forEach { index ->
-            oldArray.remove(index)
-        }
-
-        changes[basePath] = ProfileChange(oldArrayCopy, oldArray)
-    }
-
-    /**
-     * Deletes a value and records the change.
-     */
-    private fun deleteValue(
-        parent: JSONObject,
-        key: String,
-        value: Any,
-        path: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        if (value is JSONObject) {
-            recordAllLeafDeletions(value, path, changes)
-        } else {
-            changes[path] = ProfileChange(value, null)
-        }
-        parent.remove(key)
-    }
-
-    /**
-     * Handles non-delete merge operations.
-     */
-    private fun handleMerge(
-        target: JSONObject,
-        key: String,
-        newValue: Any,
-        currentPath: String,
-        changes: MutableMap<String, ProfileChange>,
-        operation: MergeOperation
-    ) {
-        if (!target.has(key)) {
-            handleMissingKey(target, key, newValue, currentPath, changes, operation)
-            return
-        }
-
-        val oldValue = target.get(key)
-
-        when {
-            oldValue is JSONObject && newValue is JSONObject -> {
-                // Recurse into nested objects
-                mergeRecursive(oldValue, newValue, currentPath, changes, operation)
-            }
-            oldValue is JSONArray && newValue is JSONArray -> {
-                // Handle array operations
-                handleArrayMerge(target, key, oldValue, newValue, currentPath, changes, operation)
-            }
-            oldValue is Number && newValue is Number &&
-                    operation in listOf(MergeOperation.INCREMENT, MergeOperation.DECREMENT) -> {
-                // Handle arithmetic operations
-                handleNumberOperation(target, key, oldValue, newValue, currentPath, changes, operation)
-            }
-            else -> {
-                // Handle simple value update
-                handleValueUpdate(target, key, oldValue, newValue, currentPath, changes)
-            }
-        }
-    }
-
-    /**
-     * Handles keys that don't exist in the target.
-     */
-    private fun handleMissingKey(
-        target: JSONObject,
-        key: String,
-        newValue: Any,
-        currentPath: String,
-        changes: MutableMap<String, ProfileChange>,
-        operation: MergeOperation
-    ) {
-        // Skip adding keys for arithmetic operations
-        if (operation in listOf(MergeOperation.INCREMENT, MergeOperation.DECREMENT)) {
-            return
-        }
-
-        target.put(key, newValue)
-
-        if (newValue is JSONObject) {
-            recordAllLeafValues(newValue, currentPath, changes)
-        } else {
-            changes[currentPath] = ProfileChange(null, newValue)
-        }
-    }
-
-    /**
-     * Handles array merge operations.
-     */
-    private fun handleArrayMerge(
-        parentJson: JSONObject,
-        key: String,
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        currentPath: String,
-        changes: MutableMap<String, ProfileChange>,
-        operation: MergeOperation
-    ) {
-        if (newArray.length() == 0) return
-
-        when (operation) {
-            MergeOperation.ARRAY_ADD -> handleArrayAdd(oldArray, newArray, currentPath, changes)
-            MergeOperation.ARRAY_REMOVE -> handleArrayRemove(parentJson, key, oldArray, newArray, currentPath, changes)
-            MergeOperation.UPDATE, MergeOperation.INCREMENT, MergeOperation.DECREMENT -> {
-                if (shouldMergeArrayElements(newArray)) {
-                    mergeArrayElements(oldArray, newArray, currentPath, changes, operation)
-                } else {
-                    handleArrayReplacement(parentJson, key, oldArray, newArray, currentPath, changes)
-                }
-            }
-            else -> {}
-        }
-    }
-
-    /**
-     * Determines if array should be merged element-wise or replaced.
-     */
-    private fun shouldMergeArrayElements(array: JSONArray): Boolean {
-        return (0 until array.length()).any { i -> array.opt(i) is JSONObject }
-    }
-
-    /**
-     * Merges array elements individually.
-     */
-    private fun mergeArrayElements(
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        basePath: String,
-        changes: MutableMap<String, ProfileChange>,
-        operation: MergeOperation
-    ) {
-        val oldArrayCopy = copyArray(oldArray)
-        var arrayModified = false
-
-        for (i in 0 until newArray.length()) {
-            if (i >= oldArray.length()) {
-                handleOutOfBoundsIndex(oldArray, newArray, i, operation)
-                arrayModified = true
-                continue
-            }
-
-            val oldElement = oldArray.get(i)
-            val newElement = newArray.get(i)
-
-            when {
-                oldElement is JSONObject && newElement is JSONObject -> {
-                    val elementChanges = mutableMapOf<String, ProfileChange>()
-                    mergeRecursive(oldElement, newElement, "", elementChanges, operation)
-                    if (elementChanges.isNotEmpty()) {
-                        arrayModified = true
+                MergeOperation.DELETE -> {
+                    deleteHandler.handleDelete(
+                        target, key, newValue, currentPath, changes
+                    ) { target, source, path, changes ->
+                        mergeRecursive(target, source, path, changes, MergeOperation.DELETE)
                     }
                 }
-                oldElement is Number && newElement is Number -> {
-                    val result = applyNumberOperation(oldElement, newElement, operation)
-                    if (!areEqual(oldElement, result)) {
-                        oldArray.put(i, result)
-                        arrayModified = true
+                else -> {
+                    updateHandler.handleMerge(
+                        target, key, newValue, currentPath, changes, operation
+                    ) { target, source, path, changes ->
+                        mergeRecursive(target, source, path, changes, operation)
                     }
                 }
-                operation == MergeOperation.UPDATE && !areEqual(oldElement, newElement) -> {
-                    oldArray.put(i, newElement)
-                    arrayModified = true
-                }
-            }
-        }
-
-        if (arrayModified) {
-            changes[basePath] = ProfileChange(oldArrayCopy, oldArray)
-        }
-    }
-
-    /**
-     * Handles array indices that are out of bounds.
-     */
-    private fun handleOutOfBoundsIndex(
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        index: Int,
-        operation: MergeOperation
-    ) {
-        if (operation != MergeOperation.UPDATE) return
-
-        val newElement = newArray.get(index)
-        while (oldArray.length() <= index) {
-            oldArray.put(JSONObject.NULL)
-        }
-        oldArray.put(index, newElement)
-    }
-
-    /**
-     * Adds string values to array (allows duplicates).
-     * If a string starts with $D_ prefix, removes it and converts to long.
-     */
-    private fun handleArrayAdd(
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        path: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val oldArrayCopy = copyArray(oldArray)
-        var modified = false
-
-        for (i in 0 until newArray.length()) {
-            val item = newArray.get(i)
-            if (item is String) {
-                val processedItem = processDatePrefix(item)
-                oldArray.put(processedItem)
-                modified = true
-            }
-        }
-
-        if (modified) {
-            changes[path] = ProfileChange(oldArrayCopy, oldArray)
-        }
-    }
-
-    /**
-     * Removes string values from array.
-     */
-    private fun handleArrayRemove(
-        parentJson: JSONObject,
-        key: String,
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        path: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val oldArrayCopy = copyArray(oldArray)
-        val resultArray = JSONArray()
-        var modified = false
-
-        for (i in 0 until oldArray.length()) {
-            val item = oldArray.get(i)
-            if (item is String && arrayContainsString(newArray, item)) {
-                modified = true
-            } else {
-                resultArray.put(item)
-            }
-        }
-
-        if (modified) {
-            parentJson.put(key, resultArray)
-            changes[path] = ProfileChange(oldArrayCopy, resultArray)
-        }
-    }
-
-    /**
-     * Replaces entire array with new array.
-     */
-    private fun handleArrayReplacement(
-        parentJson: JSONObject,
-        key: String,
-        oldArray: JSONArray,
-        newArray: JSONArray,
-        path: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        if (!areEqual(oldArray, newArray)) {
-            parentJson.put(key, newArray)
-            changes[path] = ProfileChange(oldArray, newArray)
-        }
-    }
-
-    /**
-     * Handles arithmetic operations on numbers.
-     */
-    private fun handleNumberOperation(
-        parent: JSONObject,
-        key: String,
-        oldValue: Number,
-        newValue: Number,
-        path: String,
-        changes: MutableMap<String, ProfileChange>,
-        operation: MergeOperation
-    ) {
-        val result = applyNumberOperation(oldValue, newValue, operation)
-
-        if (!areEqual(oldValue, result)) {
-            parent.put(key, result)
-            changes[path] = ProfileChange(oldValue, result)
-        }
-    }
-
-    /**
-     * Applies arithmetic operation to numbers.
-     */
-    private fun applyNumberOperation(
-        oldValue: Number,
-        newValue: Number,
-        operation: MergeOperation
-    ): Number {
-        return when (operation) {
-            MergeOperation.INCREMENT -> addNumbers(oldValue, newValue)
-            MergeOperation.DECREMENT -> subtractNumbers(oldValue, newValue)
-            else -> oldValue
-        }
-    }
-
-    /**
-     * Handles simple value updates.
-     * If new value is a string with $D_ prefix, removes it and converts to long.
-     */
-    private fun handleValueUpdate(
-        parent: JSONObject,
-        key: String,
-        oldValue: Any,
-        newValue: Any,
-        path: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val processedOldValue = if (oldValue is String) {
-            processDatePrefix(oldValue)
-        } else {
-            oldValue
-        }
-
-        val processedNewValue = if (newValue is String) {
-            processDatePrefix(newValue)
-        } else {
-            newValue
-        }
-        
-        if (!areEqual(processedOldValue, processedNewValue)) {
-            parent.put(key, processedNewValue)
-            changes[path] = ProfileChange(processedOldValue, processedNewValue)
-        }
-    }
-
-    /**
-     * Records all leaf values in a JSONObject as additions.
-     */
-    private fun recordAllLeafValues(
-        jsonObject: JSONObject,
-        basePath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val keys = jsonObject.keys()
-        while (keys.hasNext()) {
-            val key = keys.next()
-            val value = jsonObject.get(key)
-            val currentPath = buildPath(basePath, key)
-
-            when (value) {
-                is JSONObject -> recordAllLeafValues(value, currentPath, changes)
-                else -> changes[currentPath] = ProfileChange(null, value)
             }
         }
     }
-
-    /**
-     * Records all leaf values in a JSONObject as deletions.
-     */
-    private fun recordAllLeafDeletions(
-        jsonObject: JSONObject,
-        basePath: String,
-        changes: MutableMap<String, ProfileChange>
-    ) {
-        val keys = jsonObject.keys()
-        while (keys.hasNext()) {
-            val key = keys.next()
-            val value = jsonObject.get(key)
-            val currentPath = buildPath(basePath, key)
-
-            when (value) {
-                is JSONObject -> recordAllLeafDeletions(value, currentPath, changes)
-                else -> changes[currentPath] = ProfileChange(value, null)
-            }
-        }
-    }
-
-    // ==================== Utility Methods ====================
 
     /**
      * Builds a dot-notation path from base path and key.
@@ -629,157 +111,7 @@ class ProfileStateMerger {
         return if (basePath.isEmpty()) key else "$basePath.$key"
     }
 
-    /**
-     * Adds two numbers, preserving type when possible.
-     */
-    private fun addNumbers(a: Number, b: Number): Number {
-        return when {
-            a is Int && b is Int -> a + b
-            a is Long || b is Long -> a.toLong() + b.toLong()
-            a is Float || b is Float -> a.toFloat() + b.toFloat()
-            else -> a.toDouble() + b.toDouble()
-        }
-    }
-
-    /**
-     * Subtracts two numbers, preserving type when possible.
-     */
-    private fun subtractNumbers(a: Number, b: Number): Number {
-        return when {
-            a is Int && b is Int -> a - b
-            a is Long || b is Long -> a.toLong() - b.toLong()
-            a is Float || b is Float -> a.toFloat() - b.toFloat()
-            else -> a.toDouble() - b.toDouble()
-        }
-    }
-
-    /**
-     * Deep equality check for JSON values.
-     */
-    private fun areEqual(obj1: Any?, obj2: Any?): Boolean {
-        if (obj1 == null && obj2 == null) return true
-        if (obj1 == null || obj2 == null) return false
-        if (obj1 === obj2) return true
-
-        return when {
-            obj1 is JSONObject && obj2 is JSONObject -> jsonObjectsEqual(obj1, obj2)
-            obj1 is JSONArray && obj2 is JSONArray -> jsonArraysEqual(obj1, obj2)
-            else -> obj1 == obj2
-        }
-    }
-
-    /**
-     * Checks if two JSONObjects are equal.
-     */
-    private fun jsonObjectsEqual(obj1: JSONObject, obj2: JSONObject): Boolean {
-        if (obj1.length() != obj2.length()) return false
-
-        val keys1 = obj1.keys()
-        while (keys1.hasNext()) {
-            val key = keys1.next()
-            if (!obj2.has(key)) return false
-
-            val value1 = obj1.get(key)
-            val value2 = obj2.get(key)
-
-            if (!areEqual(value1, value2)) return false
-        }
-
-        return true
-    }
-
-    /**
-     * Checks if two JSONArrays are equal.
-     */
-    private fun jsonArraysEqual(arr1: JSONArray, arr2: JSONArray): Boolean {
-        if (arr1.length() != arr2.length()) return false
-
-        for (i in 0 until arr1.length()) {
-            val value1 = arr1.get(i)
-            val value2 = arr2.get(i)
-
-            if (!areEqual(value1, value2)) return false
-        }
-
-        return true
-    }
-
-    /**
-     * Checks if array contains a specific string.
-     */
-    private fun arrayContainsString(array: JSONArray, value: String): Boolean {
-        for (i in 0 until array.length()) {
-            val item = array.get(i)
-            if (item is String && item == value) {
-                return true
-            }
-        }
-        return false
-    }
-
-    /**
-     * Creates a deep copy of a JSONArray.
-     */
-    private fun copyArray(array: JSONArray): JSONArray {
-        return JSONArray(array.toString())
-    }
-
-    /**
-     * Checks if array has elements marked for deletion.
-     */
-    private fun hasDeleteMarkerElements(array: JSONArray): Boolean {
-        return (0 until array.length()).any { i -> isDeleteMarker(array.opt(i)) }
-    }
-
-    /**
-     * Checks if array has JSONObject elements.
-     */
-    private fun hasJsonObjectElements(array: JSONArray): Boolean {
-        return (0 until array.length()).any { i ->
-            val element = array.opt(i)
-            element is JSONObject && !isDeleteMarker(element)
-        }
-    }
-
     companion object {
-        /**
-         * Special marker to indicate deletion intent.
-         * Use this value to mark fields for deletion.
-         */
-        const val DELETE_MARKER = "__CLEVERTAP_DELETE__"
-        
-        /**
-         * Prefix for date values that should be converted to long.
-         */
-        private const val DATE_PREFIX = "\$D_"
-
-        /**
-         * Checks if a value is the DELETE_MARKER.
-         */
-        @JvmStatic
-        fun isDeleteMarker(value: Any?): Boolean {
-            return value is String && value == DELETE_MARKER
-        }
-
-        /**
-         * Processes a string value, removing $D_ prefix if present and converting to long.
-         * 
-         * @param value The string value to process
-         * @return Long value if string starts with $D_ prefix, otherwise original string
-         */
-        private fun processDatePrefix(value: String): Any {
-            return if (value.startsWith(DATE_PREFIX)) {
-                try {
-                    value.removePrefix(DATE_PREFIX).toLong()
-                } catch (e: NumberFormatException) {
-                    // If conversion fails, return original string
-                    value
-                }
-            } else {
-                value
-            }
-        }
-        
         /**
          * Extension function to convert ProfileChange map to nested map format.
          * Useful for serialization or legacy compatibility.
